requirements.txt
langchain
langchain-openai
langchain-community
openai
fastapi
uvicorn[standard]
streamlit
duckduckgo-search
python-dotenv
requests
pydantic
httpx
langgraph
